JYFR:   - No decimals in driver speed.   - Path point to point elapsed time control.   - Fixed starting mode not loaded from properties.   - Second by second simulated driver control.   - Interpolation using the minimum necessary points in segments of paths.   - Fixed generation of aleatory drivers and paths.   - Minor fixes.

+++ src/main/java/es/us/lsi/hermes/smartDriver/VehicleLocation.java
@@ -8,7 +8,7 @@ public class VehicleLocation implements Serializable {
     private Double longitude;
     private Integer stress;
     private Double latitude;
-    private Double speed;
+    private int speed;
     private Integer accuracy;
 
     public String getTimeStamp() {
@@ -43,11 +43,11 @@ public class VehicleLocation implements Serializable {
         this.latitude = latitude;
     }
 
-    public Double getSpeed() {
+    public int getSpeed() {
         return speed;
     }
 
-    public void setSpeed(Double speed) {
+    public void setSpeed(int speed) {
         this.speed = speed;
     }
